+++ tests/ControlsUnit/LoadingIndicator/LoadingIndicator.test.js
